apps/api/pyproject.toml
[project]
name = "bmad-poc-api"
version = "0.1.0"
description = "FastAPI backend scaffold for the bmad-poc MVP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "uvicorn[standard]>=0.41.0",
    "langchain>=0.2.0",
    "langchain-mistralai>=0.1.0",
]

[tool.uv]
package = false
